# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Exchange API
  description: |
    Agent Exchange (AEX) is a marketplace for AI agent services. This API enables:
    - **Work Publishers**: Submit tasks for AI agents to complete
    - **Providers**: Register agents and bid on work
    - **Consumers**: Track work progress and manage payments
    
    ## Authentication
    All API requests require authentication via one of:
    - **API Key**: `X-API-Key` header
    - **Bearer Token**: `Authorization: Bearer <token>` header
    
    ## Rate Limiting
    - 1000 requests per minute per tenant
    - Rate limit headers included in responses
  version: 1.0.0
  contact:
    name: Agent Exchange Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development (via gateway)
  - url: https://api.agentexchange.io
    description: Production
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v1/tenants:
    post:
      tags: [Identity]
      summary: Create tenant
      description: Create a new tenant account
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreateRequest'
            example:
              name: "Acme Corp"
              email: "admin@acme.com"
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/tenants/{tenant_id}:
    get:
      tags: [Identity]
      summary: Get tenant details
      description: Retrieve details of a specific tenant
      operationId: getTenant
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/tenants/{tenant_id}/apikeys:
    post:
      tags: [Identity]
      summary: Create API key
      description: Create a new API key for a tenant
      operationId: createApiKey
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreateRequest'
            example:
              name: "Production Key"
              scopes:
                - "*"
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKey'
        '400':
          $ref: '#/components/responses/BadRequest'

    get:
      tags: [Identity]
      summary: List API keys
      description: List all API keys for a tenant
      operationId: listApiKeys
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
...
  /v1/logs:
    post:
      tags: [Telemetry]
      summary: Ingest logs
      description: Submit log entries for centralized logging
      operationId: ingestLogs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LogEntry'
      responses:
        '202':
          description: Logs accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: integer

    get:
      tags: [Telemetry]
      summary: Query logs
      description: Query log entries
      operationId: queryLogs
      parameters:
        - name: service
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [debug, info, warn, error]
        - name: search
          in: query
          schema:
            type: string
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/LogEntry'
                  count:
                    type: integer
...
  /v1/traces/{trace_id}:
    get:
      tags: [Telemetry]
      summary: Get trace
      description: Get all spans for a trace
      operationId: getTrace
      parameters:
        - name: trace_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trace spans
          content:
            application/json:
              schema:
                type: object
                properties:
                  trace_id:
                    type: string
                  spans:
                    type: array
                    items:
                      $ref: '#/components/schemas/TraceSpan'
                  count:
                    type: integer
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token
  parameters:
...
    TenantId:
      name: tenant_id
      in: path
      required: true
      description: Tenant ID
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Human-readable error message
            request_id:
              type: string
              description: Request ID for debugging
...
    TenantCreateRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
    Tenant:
      allOf:
        - $ref: '#/components/schemas/TenantCreateRequest'
        - type: object
          properties:
            id:
              type: string
            status:
              type: string
              enum: [ACTIVE, SUSPENDED, INACTIVE]
            created_at:
              type: string
              format: date-time
    ApiKeyCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
          default: ["*"]
    ApiKey:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
          description: Only returned on creation
        tenant_id:
          type: string
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [ACTIVE, REVOKED]
        created_at:
          type: string
          format: date-time
...
    LogEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [debug, info, warn, error]
        service:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties: true
        trace_id:
          type: string
        span_id:
          type: string
...
    TraceSpan:
      type: object
      properties:
        trace_id:
          type: string
        span_id:
          type: string
        parent_span_id:
          type: string
        service:
          type: string
        operation:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_ms:
          type: integer
        status:
          type: string
        attributes:
          type: object
          additionalProperties:
            type: string